Move `CliCommand.tsx` to `components`
