+++ web/packages/teleterm/src/ui/components/CliCommand.tsx
